move New*Command() functions to init.go for csm provider
